BIG PULL - from soma 2.1
big overhaul of everything. website color. better overall sizing and collapsing to mobile. better UI. clearer links. updated links and icons.

everything was tested ahead of time on soma 2.1. plenty more needs to be done by the goal was to match the previous version of the website but with overall updated to code arrangement/ optimization.